+++ client/src/pages/_app.js
@@ -1,4 +1,5 @@
 import "tailwindcss/tailwind.css";
+import "@/styles/globals.css";
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />;
