+++ client/src/styles/globals.css
@@ -1,16 +1,30 @@
-html,
-body {
-  padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+.loading {
+  @apply animate-loading bg-gradient-to-r from-[#1876BF] to-[#57D6FF] w-24 h-24 rounded-full;
 }
 
-a {
-  color: inherit;
-  text-decoration: none;
+.loading::before {
+  content: "";
+  display: block;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: white;
+  width: 72px;
+  height: 72px;
+  border-radius: 50%;
 }
 
-* {
-  box-sizing: border-box;
+.loading::after {
+  content: "";
+  display: block;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 100px;
+  height: 100px;
+  border: 50px solid transparent;
+  border-top: 50px solid white;
+  border-radius: 50%;
 }
